send_frame_script/pyproject.toml
[project]
name = "send-frame-script"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=12.2.0",
    "pyserial>=3.5",
]
